More docs, global_tasks convenience module
