+++ oggm/__init__.py
@@ -22,7 +22,13 @@ except ImportError:
     pass
 
 # API
-# TODO: why are some funcs here? maybe reconsider what API actually is
-from oggm.utils import entity_task, global_task, GlacierDirectory
+# Some decorators used by many
+from oggm.utils import entity_task, global_task
+
+# Classes
+from oggm.utils import GlacierDirectory
 from oggm.core.centerlines import Centerline
 from oggm.core.flowline import Flowline
+from oggm.core.flowline import FlowlineModel
+from oggm.core.flowline import FileModel
+from oggm.core.massbalance import MassBalanceModel
