+++ oggm/utils/_downloads.py
@@ -2291,72 +2291,3 @@ def get_topo_file(lon_ex=None, lat_ex=None, rgi_id=None, *,
         raise InvalidDEMError('Source: {2} no topography file available for '
                               'extent lat:{0}, lon:{1}!'.
                               format(lat_ex, lon_ex, source))
-
-
-def get_ref_mb_glaciers_candidates(rgi_version=None):
-    """Reads in the WGMS list of glaciers with available MB data.
-
-    Can be found afterwards (and extended) in cdf.DATA['RGIXX_ref_ids'].
-    """
-
-    if rgi_version is None:
-        rgi_version = cfg.PARAMS['rgi_version']
-
-    if len(rgi_version) == 2:
-        # We might change this one day
-        rgi_version = rgi_version[:1]
-
-    key = 'RGI{}0_ref_ids'.format(rgi_version)
-
-    if key not in cfg.DATA:
-        flink, _ = get_wgms_files()
-        cfg.DATA[key] = flink['RGI{}0_ID'.format(rgi_version)].tolist()
-
-    return cfg.DATA[key]
-
-
-def get_ref_mb_glaciers(gdirs, y0=None, y1=None):
-    """Get the list of glaciers we have valid mass balance measurements for.
-
-    To be valid glaciers must have more than 5 years of measurements and
-    be land terminating. Therefore, the list depends on the time period of the
-    baseline climate data and this method selects them out of a list
-    of potential candidates (`gdirs` arg).
-
-    Parameters
-    ----------
-    gdirs : list of :py:class:`oggm.GlacierDirectory` objects
-        list of glaciers to check for valid reference mass balance data
-    y0 : int
-        override the default behavior which is to check the available
-        climate data (or PARAMS['ref_mb_valid_window']) and decide
-    y1 : int
-        override the default behavior which is to check the available
-        climate data (or PARAMS['ref_mb_valid_window']) and decide
-
-    Returns
-    -------
-    ref_gdirs : list of :py:class:`oggm.GlacierDirectory` objects
-        list of those glaciers with valid reference mass balance data
-
-    See Also
-    --------
-    get_ref_mb_glaciers_candidates
-    """
-
-    # Get the links
-    ref_ids = get_ref_mb_glaciers_candidates(gdirs[0].rgi_version)
-
-    # We remove tidewater glaciers and glaciers with < 5 years
-    ref_gdirs = []
-    for g in gdirs:
-        if g.rgi_id not in ref_ids or g.is_tidewater:
-            continue
-        try:
-            mbdf = g.get_ref_mb_data(y0=y0, y1=y1)
-            if len(mbdf) >= 5:
-                ref_gdirs.append(g)
-        except RuntimeError as e:
-            if 'Please process some climate data before call' in str(e):
-                raise
-    return ref_gdirs
